volatility-engine/requirements.txt
Flask
gunicorn
google-cloud-storage
google-cloud-core
volatility3
